Integrate YAML-configurable pipeline into martinize2
