+++ vermouth/tests/test_cli_builder.py
@@ -6,7 +6,7 @@ import sys
 from pathlib import Path
 sys.path.insert(0, str(Path(__file__).parent))
 import pytest
-from api import CLIBuilder
+from vermouth.pipeline import CLIBuilder
