+++ vermouth/tests/test_hypotest_clibuilder.py
@@ -7,7 +7,7 @@ import pytest
 
 sys.path.insert(0, str(Path(__file__).resolve().parent))
 
-from api import CLIBuilder
+from vermouth.pipeline import CLIBuilder
 
 
 @st.composite
