+++ vermouth/tests/test_hypotest_pipelinebuilder.py
@@ -8,7 +8,7 @@ import types
 
 sys.path.insert(0, str(Path(__file__).resolve().parent))
 
-from api import PipelineBuilder
+from vermouth.pipeline import PipelineBuilder
 from vermouth.processors.processor import Pipeline, Processor
